tests/demo/requirements.txt
six
SQLAlchemy
uliweb3
uliweb-apijson
